A64: Fix compiler warning in r19444.
BUG=
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/167463004

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@19447 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/a64/macro-assembler-a64.cc
@@ -598,7 +598,8 @@ void MacroAssembler::EmitVeneers(bool need_protection) {
         branch->SetImmPCOffsetTarget(veneer);
         b(label);
 #ifdef DEBUG
-        ASSERT(SizeOfCodeGeneratedSince(&size_check) <= kMaxVeneerCodeSize);
+        ASSERT(SizeOfCodeGeneratedSince(&size_check) <=
+               static_cast<uint64_t>(kMaxVeneerCodeSize));
         size_check.Unuse();
 #endif
